+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## Version 15.2.0
+_28 January 2019_
+
+### Table Modifications
+* Restrict participant unique id and test task id to be 20 characters
+
+### Data Modifications
+* Add ROLE_ONC to the user_permission table and convert all existing users except Ai's admin account from ROLE_ADMIN to ROLE_ONC.
+* Update soft delete triggers to include recently added references to certified products.
+* Remove ROLE_ONC_STAFF permission and convert existing ROLE_ONC_STAFF user accounts to be ROLE_ONC.
+
+---
+
 ## Version 15.1.0
 _14 January 2019_